feat : make greeting page
