+++ src/app/common/header/header.tsx
@@ -1,13 +1,34 @@
 "use client";
 
 import Image from 'next/image'
-import useScrollDirection from "@/app/common/use-scroll-direction/useScrollDirection";
+
+import {useEffect, useState} from "react";
+
+export function useScrollDirection() {
+    const [scrollDirection, setScrollDirection] = useState('');
+    useEffect(() => {
+        let lastScrollY = window.pageYOffset;
+        const updateScrollDirection = () => {
+            const scrollY = window.pageYOffset;
+            const direction = scrollY > lastScrollY ? 'down' : 'up';
+            if (direction !== scrollDirection) {
+                setScrollDirection(direction)
+            }
+            lastScrollY = scrollY > 0 ? scrollY : 0;
+        };
+        window.addEventListener("scroll", updateScrollDirection);
+        return () => {
+            window.removeEventListener("scroll", updateScrollDirection); // clean up
+        }
+    }, [scrollDirection])
+    return scrollDirection;
+};
 
 export default function Header() {
     const scrollDirection = useScrollDirection();
     return (
         <main
-            className={`sticky ${scrollDirection === 'down' ? '-top-[80px]' : 'top-0'} bg-white py-[22px] px-[5%] h-[80px] transition-all duration-500`}>
+            className={`sticky ${scrollDirection === 'down' ? '-top-[80px]' : 'top-0'} bg-white py-[22px] px-[5%] h-[80px] transition-all duration-300`}>
             <nav className='flex justify-between'>
                 <div className='flex w-[50%] items-center gap-[8px]'>
                     <Image
