--- src/app/common/use-scroll-direction/useScrollDirection.tsx
@@ -1,23 +1,0 @@
-"use client";
-
-import {useEffect, useState} from "react";
-
-export default function useScrollDirection() {
-    const [scrollDirection, setScrollDirection] = useState('');
-    useEffect(() => {
-        let lastScrollY = window.pageYOffset;
-        const updateScrollDirection = () => {
-            const scrollY = window.pageYOffset;
-            const direction = scrollY > lastScrollY ? 'down' : 'up';
-            if (direction !== scrollDirection) {
-                setScrollDirection(direction)
-            }
-            lastScrollY = scrollY > 0 ? scrollY : 0;
-        };
-        window.addEventListener("scroll", updateScrollDirection);
-        return () => {
-            window.removeEventListener("scroll", updateScrollDirection); // clean up
-        }
-    }, [scrollDirection])
-    return scrollDirection;
-};
