+++ src/app/page.tsx
@@ -1,9 +1,9 @@
-import Image from 'next/image'
+import Greeting from "@/app/component/greeting/greeting";
 
 export default function Home() {
   return (
-    <main className="flex min-h-screen flex-col items-center justify-between p-24 bg-black h-[2000px]">
-      <h1 className={"text-white"}>HELLO WORLD</h1>
+    <main className="flex flex-col items-center bg-white">
+        <Greeting/>
     </main>
   )
 }
